% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{covid_br_states}
\alias{covid_br_states}
\title{Corona virus cases for each brazilian state.}
\format{A data frame with six variables:
\itemize{
\item{\strong{date}} {The date (yyyy-mm-dd)}
\item{\strong{state}} {The name of the state}
\item{\strong{state_ID}} {the code of the state}
\item{\strong{suspected_cases}} {The official number of suspected cases on a given date}
\item{\strong{confirmed_cases}} {The official number of confirmed cases on a given date}
\item{\strong{not_confirmed_cases}} {The official number of discarded cases on a given date}
\item{\strong{deaths}} {The official number of deaths by covid-19}
}}
\source{
Data gathered from:
\url{https://saude.gov.br/}
}
\usage{
covid_br_states
}
\description{
Corona virus cases for each brazilian state.
}
\keyword{datasets}
