% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{covid_global}
\alias{covid_global}
\title{Historical number of covid-19 cases for the world.}
\format{A data frame with four variables:
\itemize{
\item{\strong{state}} {The name of the state (country)}
\item{\strong{pop}} {Total number of confirmed covid-19 cases}
\item{\strong{pop}} {Total number of confirmed covid-19 deaths}
\item{\strong{date}} {The date (yyyy-mm-dd)}
}}
\usage{
covid_global
}
\description{
Historical number of covid-19 cases for the world.
}
\keyword{datasets}
