% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{covid_regions}
\alias{covid_regions}
\title{Historical number of covid-19 cases for regions in Brazil.}
\format{A data frame with nine variables:
\itemize{
\item{\strong{region}} {The region of Brazil}
\item{\strong{state}} {The name of the state}
\item{\strong{state_ID}} {the code of the state}
\item{\strong{state_uid}} {the uid code of the state}
\item{\strong{suspected_cases}} {The official number of suspected cases on a given date}
\item{\strong{confirmed_cases}} {The official number of confirmed cases on a given date}
\item{\strong{not_confirmed_cases}} {The official number of discarded cases on a given date}
\item{\strong{deaths}} {The official number of deaths by covid-19}
\item{\strong{date}} {The date (yyyy-mm-dd)}
}}
\source{
Data downloaded from: \url{http://plataforma.saude.gov.br/novocoronavirus/} with the
package \code{coronabr}: \url{https://github.com/liibre/coronabr}.
}
\usage{
covid_regions
}
\description{
Historical number of covid-19 cases for regions in Brazil.
}
\keyword{datasets}
