% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadCovid19}
\alias{downloadCovid19}
\alias{downloadCovid}
\title{Function to download COVID-19 data from web repositories}
\usage{
downloadCovid19(from = c("brazil", "world"))
}
\arguments{
\item{from}{label associated with the repository from which the data will be downloaded: "brazil" for the official's brazilian repository, "brazil2" for the non-official repository provided by Brasil.io, and "world" for the Johns Hopkins University's repository.}
}
\value{
tibble/data.frame containing the downloaded data.
}
\description{
This function downloads the pandemic COVID-19 data from two different repositories: the official Brazilian's repository mantained by the Brazilian Government (https://covid.saude.gov.br), which contains data of the pandemic in Brazil at country/state/region/city levels, and from the John Hopkins University's repository (https://github.com/CSSEGISandData/COVID-19), which has been widely used all over the world as a reliable source of data information on the COVID-19 pandemic at a global level (countries and territories).
}
\examples{
\donttest{
library(covid19br)

# Downloading Brazilian COVID-19 data:
brazil <- downloadCovid19(from = "brazil")

# Downloading world COVID-19 data:
world <- downloadCovid19(from = "world")
}

}
