% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{election2018States}
\alias{election2018States}
\title{Results of the 2018 presidential election in Brazil by state.}
\format{
A data frame with 27 rows and 5 variables:
\itemize{
  \item region: regions' names.
  \item state: states' names.
  \item Bolsonaro: count of votes obtained by the President-elected Jair Bolosnaro.
  \item Haddad: count of votes obtained by the defeated candidate Fernando Haddad.
  \item pop: estimated population in 2019.
}
}
\source{
Tribunal Superior Eleitoral (TSE). URL:  \url{https://www.tse.jus.br/eleicoes/estatisticas}.
}
\description{
Dataset containing the results of the 2018 presidential election in Brazil.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
