% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_geo.R
\docType{data}
\name{geoworld}
\alias{geoworld}
\title{World-level georeferenced data}
\format{
A data frame with 241 rows and 11 variables:
\itemize{
  \item country: country's name.
  \item continent: continent's name.
  \item region: regions' names.
  \item subregion: subregion's name.
  \item pop: estimated population.
  \item pais: country's name in Portuguese.
  \item country_code: numerical code attributed to countries.
  \item continent_code: numerical code attributed to continents.
  \item region_code: numerical code attributed to regions.
  \item subregion_code: numerical code attributed to subregions.
  \item geometry: georeferenced data needed to plot maps.
}
}
\source{
\itemize{
    \item World map: \url{https://CRAN.R-project.org/package=rnaturalearthdata}
  }
}
\description{
Data set containing the world population and geographical information on country level.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
