% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ibgeCities}
\alias{ibgeCities}
\title{City-level georeferenced data}
\format{
A data frame with 5570 rows and 10 variables:
\itemize{
  \item region: regions' names
  \item state: states' names.
  \item city: cities' names.
  \item pop: estimated population in 2019.
  \item region_code: numerical code attributed to regions
  \item state_code: numerical code attributed to states
  \item mesoregion_code: numerical code attributed to mesoregions
  \item microregion_code: numerical code attributed to microregions
  \item city_code: numerical code attributed to cities
  \item geometry: georeferenced data needed to plot maps.
  \item area: area (in Km^2) of the brazilian cities
  \item demoDens: demographic density of the brazilian cities.
}
}
\source{
Instituto Brasileiro de Geografia e Estatística (IBGE):
  \itemize{
    \item Shapefiles: \url{https://www.ibge.gov.br/geociencias/downloads-geociencias.html}
    \item Population: \url{https://www.ibge.gov.br/estatisticas/sociais/populacao/9103-estimativas-de-populacao.html?=&t=resultados}
  }
}
\description{
Data set obtaind from the Instituto Brasileiro de Geografia e Estatística (IBGE) with data on the Brazilian population and geographical information on city level.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
