% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{info_states}
\alias{info_states}
\title{Demographic information for Brazilian states}
\format{A data frame with four variables:
\itemize{
\item{\strong{date}} {The date (yyyy-mm-dd)}
\item{\strong{state}} {The name of the state}
\item{\strong{pop}} {State population}
\item{\strong{area}} {State area in km^2}
\item{\strong{density}} {State density in people per km2}
}}
\usage{
info_states
}
\description{
Demographic information for Brazilian states
}
\keyword{datasets}
