% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ipeaCities}
\alias{ipeaCities}
\title{Brazilian development human indexes by cities}
\format{
A data frame with 5570 rows and 9 variables:
\itemize{
  \item region: regions' names.
  \item state: states' names.
  \item city: states' names.
  \item city_code: numerical code attributed to cities
  \item DHI: average development human index.
  \item EDHI: average educational development human index.
  \item LDHI: average longevity development human index.
  \item IDHI: average income development human index.
  \item pop: estimated population in 2019.
}
}
\source{
Instituto de Pesquisa Econômica Aplicada (IPEA). URL: \url{https://www.ipea.gov.br/ipeageo/bases.html}.
}
\description{
Dataset on the development humam indexes provided the Instituto de Pesquisa Econômica Aplicada in 2010.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
