% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ipeaRegions}
\alias{ipeaRegions}
\title{Brazilian development human indexes by regions}
\format{
A data frame with 5 rows and 6 variables:
\itemize{
  \item region: regions' names.
  \item DHI: development human index.
  \item EDHI: educational development human index.
  \item LDHI: longevity development human index.
  \item IDHI: income development human index.
  \item pop: estimated population in 2019.
}
}
\source{
Instituto de Pesquisa Econômica Aplicada (IPEA). URL: \url{https://www.ipea.gov.br/ipeageo/bases.html}.
}
\description{
Dataset on the development humam indexes provided the Instituto de Pesquisa Econômica Aplicada in 2010.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
