% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{map_regions}
\alias{map_regions}
\title{Brazilian map with regions}
\format{Simple feature collection with 27 features and 5 fieldS}
\source{
generated with
library(geobr)
map_regions <- read_region(year=2018)
}
\usage{
map_regions
}
\description{
Brazilian map with regions
}
\keyword{datasets}
