% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{map_states}
\alias{map_states}
\title{Brazilian map with states}
\format{Simple feature collection with 27 features and 5 fieldS}
\source{
generated with
library(geobr)
states <- read_state(year=2018)
}
\usage{
map_states
}
\description{
Brazilian map with states
}
\keyword{datasets}
