% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_geo}
\alias{add_geo}
\title{Adds georeferenced data to the downloaded data}
\usage{
add_geo(data, ...)
}
\arguments{
\item{data}{a data set downloaded using the covid19br::downloadCovid19() function.}

\item{...}{further arguments passed to other methods.}
}
\value{
the data set with the added georeferenced data.
}
\description{
Adds georeferenced data to the downloaded data
}
