% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{package}
\name{covid10br-package}
\alias{covid10br-package}
\title{Corona virus (covid-19) in Brasil.}
\description{
Updated data on the number of cases of corona virus in Brasil.
}
\details{
This package is build to keep an updated repository with the oficial data on corona virus for Brasil.
The source of data source is: Ministério da Saúde (https://saude.gov.br/).
}
\section{Bug reports}{


\url{https://github.com/paternogbc/covid19br/issues}.
}

\section{How to contribute?}{


You can find instructions on how to contribute to \bold{teads} at
this link: \url{https://github.com/paternogbc/covid19br}
}

\author{
Gustavo Brant Paterno
}
