% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadCovid19}
\alias{downloadCovid19}
\alias{downloadCovid}
\title{Function to download COVID-19 data from web repositories}
\usage{
downloadCovid19(from = c("brazil", "world"))
}
\arguments{
\item{from}{repository from which the data will be downloaded: "brazil" (default) "world".}
}
\value{
when the Brazilian repository is chosen the function returns a list containing the downloaded COVID-19 Brazilian data organized by country, region, state, and city levels; otherwise, the function returns a tibble with the downloaded data at the world-level.
}
\description{
This function downloads the pandemic COVID-19 data from two different repositories: the official Brazilian's repository maintained by the Brazilian Government (https://covid.saude.gov.br), which contains data of the pandemic in Brazil at country/state/region/city levels, and from the John Hopkins University's repository (https://github.com/CSSEGISandData/COVID-19), which has been widely used around the world as a reliable source of data information on the COVID-19 pandemic at a global level.
}
\examples{
\donttest{
library(covid19br)

# Downloading Brazilian COVID-19 data:
brazil <- downloadCovid19(from = "brazil")

# Downloading world COVID-19 data:
world <- downloadCovid19(from = "world")
}

}
